bugfix for BrozzlerWorker._needs_browsing
I'm sorry but with my previous commit I introduced a bug in
``BrozzlerWorker._needs_browsing`` method.

More specifically, if the ``brozzler_spy`` param is False (this happens
when ``youtube_dl`` is disabled), ``_needs_browsing`` method returns
always ``False`` and this messes with ``brozzle_page`` workflow. The
browser is nevenr visiting the page.

I'm fixing this here.