Browser.configure_browser optimisation
We use `Network.setExtraHTTPHeaders` and then we use `self._wait_for` to wait
for the result.
This is redundant as we don't do anything with the result, we shouldn't
wait here.

Also, in the same method we use `Network.setUserAgentOverride` and we
don't wait for that. I'm saying this to indicate that waiting isn't
necessary after setting browser options like these.