Add headless chrome option
Using `--headless` chromium-browser option, we run all Brozzler without
requiring an X display.
https://chromium.googlesource.com/chromium/src/+/lkgr/headless/README.md

Benchmarks are ~20% faster using `--headless`.

This option was introduced in chromium-browser 59 but was unstable. I've
tested it with v73.0.3683.86 and its running well.

With this PR, we add it as an option that is disabled by default.