Make brozzler.behaviors more configurable
`behaviors.yaml` is no longer hard-coded but its an option and
`behaviors.yaml` is the default value. This way, we could use another
YAML conf to check for different set of behaviors if necessary.

This is required because we need to define login behaviors as a
different set and we'll reuse the same method to load them.