Define new login behavior type
Define new `login-behaviors.yaml` which are JS behaviors we run when the
target page is behind a paywall and we need to go to another page to
login.

The problem with the current system is that the login form needs to be
in the target URL. If there isn't a login form, Brozzler cannot login.
We try to fix this issue.

As a first example, we use www.thetimes.co.uk which is using an approach
like that. It is using a paywall and you need to go to
account.thetimes.co.uk to login.

To implement this, we reuse `brozzler.behavior_script` and `brozzler.behaviors`.
We run the new behavior just before `try_login`.
Our aim is to go to the login page of the target site before running
`try_login`.