Optimize check_version
Every time we create a new `Chrome` instance, we run `check_version`
which executes a subprocess that runs `chrome_exe --version`.

Chrome version changes very rarely and when it does, we restart Brozzler anyway.

Since we create a new `Chrome` instance every time we run a fresh
`Browser` (which may be a lot), it makes sense to cache `check_version`.