behaviors: add loop control selector
Right now, our loop config in behaviours is purely based on the loop
limit. We can be smarter than that, though - sometimes we can know
when it's safe to *stop* clicking the same element over and over by
running another selector.

A good example of this is page pagination. Most sites with
JavaScript-based pagination will also disable the "next" button once
the browser has reached the last page of results. Using that, we can
configure a behaviour that looks something like this:

```yaml
- url_regex: '^.*$'
  behavior_js_template: umbraBehavior.js.j2
  request_idle_timeout_sec: 10
  default_parameters:
    interval: 10
    actions:
      - selector: .next
        repeatSameElement: true
        repeatUntilSelector: .next[disabled]
        # the repeatUntil selector makes long loops safe
        # since we'll break out of the loop once we hit the
        # disabled button!
        limit: 1000
```