behaviors: optionally collect outlinks while looping
This solves one of the existing limitations of our outlink gathering
system: we run it once, after all behaviours have completed. For
some interactive pages, particularly single-page apps with paginated
data, it means that we'll completely miss that content since it won't
be in the DOM anymore by the time we get around to it.

In a previous PR, #433, I made the outlink gathering logic reusable
by ensuring it's possible for us to dynamically call the outlink
gathering function at any time. In addition, in #429, I made it
possible for behaviours to return outlinks to Brozzler; if the
behaviour chooses to return outlinks, then Brozzler will add them to
the set it extracts after behavours complete.

This branch uses both of those by introducing new functionality in
behaviours. We always inject the outlink gathering code before running
behaviours, so they can now run it at will. It also introduces two new
behaviour parameters:

* `extractOutlinks` - If set to `true`, then the behaviour script will
  call the outlink gathering logic and return any outlinks it harvested
  to Brozzler. Defaults to `false`.
* `extractOutlinksInLoop` - If set to `true`, then the behaviour script
  will gather outlinks every iteration of the loop. This combines great
  with `repeatSameElement`, since it means the behaviour script can
  click a `next` pagination button and then immediately gather whatever
  new outlinks have appeared on the page. Defaults to `false`.